test for triangle angle

+++ PolygonChecker/GetTriangleAngle.c
@@ -4,9 +4,9 @@
 #define PI 3.14159265358979323846
 double* getTriangleAngles(int* triangleSides, double* triangleAngles)
 {
-	    double a = triangleSides[0];
-		double b = triangleSides[1];
-		double c = triangleSides[2];
+	    int a = triangleSides[0];
+		int b = triangleSides[1];
+		int c = triangleSides[2];
 		double cosA = (b * b + c * c - a * a) / (2 * b * c);
 		double cosB = (a * a + c * c - b * b) / (2 * a * c);
 		double cosC = (a * a + b * b - c * c) / (2 * a * b);
